sfs-361 refactor tests

+++ shared/annual_billing_information.go
@@ -2,9 +2,9 @@ package shared
 
 type AnnualBillingInformation struct {
 	AnnualBillingYear        string
-	DemandedExpectedCount    int64
-	DemandedIssuedCount      int64
-	DemandedSkippedCount     int64
-	DirectDebitExpectedCount int64
-	DirectDebitIssuedCount   int64
+	DemandedExpectedCount    int
+	DemandedIssuedCount      int
+	DemandedSkippedCount     int
+	DirectDebitExpectedCount int
+	DirectDebitIssuedCount   int
 }
